+++ packages/terra-core-docs/src/terra-dev-site/doc/hyperlink/example/UnderlineHiddenHyperlink.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import Hyperlink from 'terra-hyperlink';
-
-export default () => <Hyperlink href="https://www.cerner.com" isUnderlineHidden>Underline hidden hyperlink</Hyperlink>;