+++ packages/terra-core-docs/src/terra-dev-site/doc/hyperlink/example/VideoHyperlink.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
 import Hyperlink from 'terra-hyperlink';
 
-export default () => <Hyperlink href="https://www.cerner.com" variant="video">Video hyperlink</Hyperlink>;
+export default () => <Hyperlink href="https://www.cerner.com" variant="video" text="Video hyperlink" />;