+++ packages/terra-core-docs/src/terra-dev-site/test/hyperlink/ExternalHyperlink.test.jsx
@@ -3,6 +3,6 @@ import Hyperlink from 'terra-hyperlink';
 
 export default () => (
   <div role="main">
-    <Hyperlink id="link" href="https://www.cerner.com" variant="external">External hyperlink</Hyperlink>
+    <Hyperlink id="link" href="https://www.cerner.com" variant="external" text="External hyperlink" />
   </div>
 );