+++ packages/terra-core-docs/src/terra-dev-site/test/hyperlink/ScaledHyperlink.test.jsx
@@ -7,15 +7,15 @@ const cx = classNames.bind(styles);
 
 const ScaledHyperlink = () => (
   <div role="main" className={cx('scaled-hyperlink')} id="link-example">
-    <Hyperlink href="https://www.cerner.com" variant="audio">Audio hyperlink</Hyperlink>
+    <Hyperlink href="https://www.cerner.com" variant="audio" text="Audio hyperlink" />
     <br />
-    <Hyperlink href="https://www.cerner.com" variant="document">Document hyperlink</Hyperlink>
+    <Hyperlink href="https://www.cerner.com" variant="document" text="Document hyperlink" />
     <br />
-    <Hyperlink href="https://www.cerner.com" variant="external">External hyperlink</Hyperlink>
+    <Hyperlink href="https://www.cerner.com" variant="external" text="External hyperlink" />
     <br />
-    <Hyperlink href="https://www.cerner.com" variant="image">Image hyperlink</Hyperlink>
+    <Hyperlink href="https://www.cerner.com" variant="image" text="Image hyperlink" />
     <br />
-    <Hyperlink href="https://www.cerner.com" variant="video">Video hyperlink</Hyperlink>
+    <Hyperlink href="https://www.cerner.com" variant="video" text="Video hyperlink" />
   </div>
 );