+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,10 +3,12 @@
 ## Unreleased
 
 * Added
+  * Added accessibility guide for `terra-hyperlink`
   * Added Accessibility Guide for `terra-section-header`
   * Added accessibility guide for `terra-progress-bar`
 
 * Changed
+  * Updated example for `terra-hyperlink`
   * Updated example for `terra-progress-bar`
 
 ## 1.23.0 - (March 1, 2023)